Restructure (#22)
* Seperate network_visualiser from visualiser

* - renamed pos to graph_layout
- moved graph_layout from models to network_visualiser

* refactor file name to not contain underscore

* taking the graphs classes out of models.py

* split unit tests

* added test for scenariograph.set_final_trace

* add arguments to atests

* correct orientation for bfs
