+++ atest/resources/helpers/modelgenerator.py
@@ -3,14 +3,16 @@ import string
 
 from robot.api.deco import keyword  # type:ignore
 from robotmbt.modelspace import ModelSpace
-from robotmbt.visualise.models import TraceInfo, ScenarioInfo, ScenarioGraph
+from robotmbt.visualise.models import TraceInfo, ScenarioInfo
+from robotmbt.visualise.graphs.scenariograph import ScenarioGraph
 
 
 class ModelGenerator:
     @keyword(name="Generate Trace Information")  # type: ignore
     def generate_trace_info(self, scenario_count: int) -> TraceInfo:
         """Generates a list of unique random scenarios."""
-        scenarios: list[ScenarioInfo] = ModelGenerator.generate_scenario_names(scenario_count)
+        scenarios: list[ScenarioInfo] = ModelGenerator.generate_scenario_names(
+            scenario_count)
 
         return TraceInfo(scenarios, ModelSpace())
 
@@ -28,7 +30,8 @@ class ModelGenerator:
         scen2_info: ScenarioInfo | None = trace_info.get_scenario(scen2)
 
         if scen1_info is None or scen2_info is None:
-            raise Exception(f"Ensure Scenario Follows for scenarios that did not exist! scen1={scen1}, scen2={scen2}")
+            raise Exception(
+                f"Ensure Scenario Follows for scenarios that did not exist! scen1={scen1}, scen2={scen2}")
 
         # both scenarios apparently exist, now make sure that scenario2 follows after some appearance of scenario 1:
         scen1_index: int = trace_info.trace.index(scen1_info)
