+++ atest/robotMBT tests/04__scenario_linking/01__then_to_given_linking.robot
@@ -4,7 +4,7 @@ Documentation     This test suite demonstrates direct one-on-one linking of scen
 ...               the trailing scenario.
 ...
 ...               Note that this test suite would also pass when run in Robot Framework without additional processing.
-Suite Setup       Treat this test suite Model-based
+Suite Setup       Treat this test suite Model-based  graph=scenario
 Resource          ../../resources/birthday_cards_flat.resource
 Library           robotmbt
