+++ atest/robotMBT tests/04__scenario_linking/02__swapped_then_to_given_linking.robot
@@ -5,7 +5,7 @@ Documentation     This test suite demonstrates direct one-on-one linking of scen
 ...               reverse order, this test suite would fail in a regular Robot Framework
 ...               test run. It passes when the model figures out the dependency between
 ...               the test cases and swaps their order.
-Suite Setup       Treat this test suite Model-based
+Suite Setup       Treat this test suite Model-based  graph=scenario
 Resource          ../../resources/birthday_cards_flat.resource
 Library           robotmbt
