+++ atest/robotMBT tests/05__repeating_scenarios/01__single_repetition.robot
@@ -2,7 +2,7 @@
 Documentation     This suite consists of 3 scenarios. After inserting the leading
 ...               scenario, the trailing scenario cannot be reached, unless the middle
 ...               scenario is inserted twice.
-Suite Setup       Treat this test suite Model-based
+Suite Setup       Treat this test suite Model-based  graph=scenario
 Resource          ../../resources/birthday_cards_flat.resource
 Library           robotmbt
